fix: finalize twitch send rebase follow-ups
