+++ crates/app/src/config/channels.rs
@@ -31,14 +31,11 @@ mod nostr_impl;
 mod signal_impl;
 mod twitch;
 
+pub use self::twitch::{ResolvedTwitchChannelConfig, TwitchAccountConfig, TwitchChannelConfig};
 pub use nostr_impl::{NostrAccountConfig, NostrChannelConfig, ResolvedNostrChannelConfig};
 pub(crate) use nostr_impl::{parse_nostr_private_key_hex, parse_nostr_public_key_hex};
 use signal_impl::{
     default_signal_account_env, default_signal_service_url, default_signal_service_url_env,
-};
-use self::twitch::{
-    default_twitch_access_token_env, default_twitch_api_base_url, default_twitch_oauth_base_url,
-    validate_twitch_env_pointer, validate_twitch_secret_ref_env_pointer,
 };
 
 pub(crate) const TELEGRAM_BOT_TOKEN_ENV: &str = "TELEGRAM_BOT_TOKEN";
@@ -1355,61 +1352,6 @@ impl ResolvedSignalChannelConfig {
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct TwitchAccountConfig {
-    #[serde(default)]
-    pub enabled: Option<bool>,
-    #[serde(default)]
-    pub account_id: Option<String>,
-    #[serde(default)]
-    pub access_token: Option<SecretRef>,
-    #[serde(default)]
-    pub access_token_env: Option<String>,
-    #[serde(default)]
-    pub api_base_url: Option<String>,
-    #[serde(default)]
-    pub oauth_base_url: Option<String>,
-    #[serde(default)]
-    pub channel_names: Option<Vec<String>>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ResolvedTwitchChannelConfig {
-    pub configured_account_id: String,
-    pub configured_account_label: String,
-    pub account: ChannelAccountIdentity,
-    pub enabled: bool,
-    pub access_token: Option<SecretRef>,
-    pub access_token_env: Option<String>,
-    pub api_base_url: Option<String>,
-    pub oauth_base_url: Option<String>,
-    pub channel_names: Vec<String>,
-}
-
-impl ResolvedTwitchChannelConfig {
-    pub fn access_token(&self) -> Option<String> {
-        resolve_secret_with_legacy_env(self.access_token.as_ref(), self.access_token_env.as_deref())
-    }
-
-    pub fn resolved_api_base_url(&self) -> String {
-        self.api_base_url
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_owned)
-            .unwrap_or_else(default_twitch_api_base_url)
-    }
-
-    pub fn resolved_oauth_base_url(&self) -> String {
-        self.oauth_base_url
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_owned)
-            .unwrap_or_else(default_twitch_oauth_base_url)
-    }
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
 pub struct WhatsappAccountConfig {
     #[serde(default)]
     pub enabled: Option<bool>,
@@ -1820,29 +1762,6 @@ pub struct SignalChannelConfig {
     pub service_url_env: Option<String>,
     #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
     pub accounts: BTreeMap<String, SignalAccountConfig>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(default)]
-pub struct TwitchChannelConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    #[serde(default)]
-    pub account_id: Option<String>,
-    #[serde(default)]
-    pub default_account: Option<String>,
-    #[serde(default)]
-    pub access_token: Option<SecretRef>,
-    #[serde(default = "default_twitch_access_token_env")]
-    pub access_token_env: Option<String>,
-    #[serde(default)]
-    pub api_base_url: Option<String>,
-    #[serde(default)]
-    pub oauth_base_url: Option<String>,
-    #[serde(default)]
-    pub channel_names: Vec<String>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub accounts: BTreeMap<String, TwitchAccountConfig>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
@@ -2417,22 +2336,6 @@ impl Default for SignalChannelConfig {
             signal_account_env: Some(SIGNAL_ACCOUNT_ENV.to_owned()),
             service_url: None,
             service_url_env: Some(SIGNAL_SERVICE_URL_ENV.to_owned()),
-            accounts: BTreeMap::new(),
-        }
-    }
-}
-
-impl Default for TwitchChannelConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            account_id: None,
-            default_account: None,
-            access_token: None,
-            access_token_env: Some(TWITCH_ACCESS_TOKEN_ENV.to_owned()),
-            api_base_url: Some(default_twitch_api_base_url()),
-            oauth_base_url: Some(default_twitch_oauth_base_url()),
-            channel_names: Vec::new(),
             accounts: BTreeMap::new(),
         }
     }
@@ -5878,190 +5781,6 @@ impl SignalChannelConfig {
     }
 }
 
-impl TwitchChannelConfig {
-    pub(crate) fn validate(&self) -> Vec<ConfigValidationIssue> {
-        let mut issues = Vec::new();
-        validate_channel_account_integrity(
-            &mut issues,
-            "twitch",
-            self.default_account.as_deref(),
-            self.accounts.keys(),
-        );
-        validate_twitch_env_pointer(
-            &mut issues,
-            "twitch.access_token_env",
-            self.access_token_env.as_deref(),
-            "twitch.access_token",
-        );
-        validate_twitch_secret_ref_env_pointer(
-            &mut issues,
-            "twitch.access_token",
-            self.access_token.as_ref(),
-        );
-        for (raw_account_id, account) in &self.accounts {
-            let account_id = normalize_channel_account_id(raw_account_id);
-            let access_token_field_path = format!("twitch.accounts.{account_id}.access_token");
-            let access_token_env_field_path = format!("{access_token_field_path}_env");
-            validate_twitch_env_pointer(
-                &mut issues,
-                access_token_env_field_path.as_str(),
-                account.access_token_env.as_deref(),
-                access_token_field_path.as_str(),
-            );
-            validate_twitch_secret_ref_env_pointer(
-                &mut issues,
-                access_token_field_path.as_str(),
-                account.access_token.as_ref(),
-            );
-        }
-        issues
-    }
-
-    pub fn access_token(&self) -> Option<String> {
-        resolve_secret_with_legacy_env(self.access_token.as_ref(), self.access_token_env.as_deref())
-    }
-
-    pub fn resolved_api_base_url(&self) -> String {
-        self.api_base_url
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_owned)
-            .unwrap_or_else(default_twitch_api_base_url)
-    }
-
-    pub fn resolved_oauth_base_url(&self) -> String {
-        self.oauth_base_url
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_owned)
-            .unwrap_or_else(default_twitch_oauth_base_url)
-    }
-
-    pub fn configured_account_ids(&self) -> Vec<String> {
-        let configured_ids = configured_account_ids(self.accounts.keys());
-        if configured_ids.is_empty() {
-            return vec![self.default_configured_account_id()];
-        }
-        configured_ids
-    }
-
-    pub fn default_configured_account_selection(&self) -> ChannelDefaultAccountSelection {
-        resolve_default_configured_account_selection(
-            self.accounts.keys(),
-            self.default_account.as_deref(),
-            self.resolved_account_identity().id.as_str(),
-        )
-    }
-
-    pub fn default_configured_account_id(&self) -> String {
-        self.default_configured_account_selection().id
-    }
-
-    pub fn resolved_account_route(
-        &self,
-        requested_account_id: Option<&str>,
-        selected_configured_account_id: &str,
-    ) -> ChannelResolvedAccountRoute {
-        resolve_channel_account_route(
-            self.accounts.keys(),
-            self.default_account.as_deref(),
-            self.resolved_account_identity().id.as_str(),
-            requested_account_id,
-            selected_configured_account_id,
-        )
-    }
-
-    pub fn resolve_account(
-        &self,
-        requested_account_id: Option<&str>,
-    ) -> CliResult<ResolvedTwitchChannelConfig> {
-        let configured = self.resolve_configured_account_selection(requested_account_id)?;
-        let account_override = configured
-            .account_key
-            .as_deref()
-            .and_then(|key| self.accounts.get(key));
-
-        let merged = TwitchChannelConfig {
-            enabled: self.enabled
-                && account_override
-                    .and_then(|account| account.enabled)
-                    .unwrap_or(true),
-            account_id: account_override
-                .and_then(|account| account.account_id.clone())
-                .or_else(|| self.account_id.clone()),
-            default_account: None,
-            access_token: account_override
-                .and_then(|account| account.access_token.clone())
-                .or_else(|| self.access_token.clone()),
-            access_token_env: account_override
-                .and_then(|account| account.access_token_env.clone())
-                .or_else(|| self.access_token_env.clone()),
-            api_base_url: account_override
-                .and_then(|account| account.api_base_url.clone())
-                .or_else(|| self.api_base_url.clone()),
-            oauth_base_url: account_override
-                .and_then(|account| account.oauth_base_url.clone())
-                .or_else(|| self.oauth_base_url.clone()),
-            channel_names: account_override
-                .and_then(|account| account.channel_names.clone())
-                .unwrap_or_else(|| self.channel_names.clone()),
-            accounts: BTreeMap::new(),
-        };
-        let account = merged.resolved_account_identity();
-
-        Ok(ResolvedTwitchChannelConfig {
-            configured_account_id: configured.id,
-            configured_account_label: configured.label,
-            account,
-            enabled: merged.enabled,
-            access_token: merged.access_token,
-            access_token_env: merged.access_token_env,
-            api_base_url: merged.api_base_url,
-            oauth_base_url: merged.oauth_base_url,
-            channel_names: merged.channel_names,
-        })
-    }
-
-    pub fn resolve_account_for_session_account_id(
-        &self,
-        session_account_id: Option<&str>,
-    ) -> CliResult<ResolvedTwitchChannelConfig> {
-        resolve_account_for_session_account_id(
-            session_account_id,
-            || self.resolve_account(session_account_id),
-            || self.configured_account_ids(),
-            |configured_id| self.resolve_account(Some(configured_id)),
-            |resolved| resolved.account.id.as_str(),
-        )
-    }
-
-    pub fn resolved_account_identity(&self) -> ChannelAccountIdentity {
-        if let Some((id, label)) = resolve_configured_account_identity(self.account_id.as_deref()) {
-            return ChannelAccountIdentity {
-                id,
-                label,
-                source: ChannelAccountIdentitySource::Configured,
-            };
-        }
-
-        default_channel_account_identity()
-    }
-
-    fn resolve_configured_account_selection(
-        &self,
-        requested_account_id: Option<&str>,
-    ) -> CliResult<ResolvedConfiguredAccount> {
-        resolve_configured_account_selection(
-            self.accounts.keys(),
-            requested_account_id,
-            self.default_account.as_deref(),
-            self.resolved_account_identity().id.as_str(),
-        )
-    }
-}
-
 impl WhatsappChannelConfig {
     pub(crate) fn validate(&self) -> Vec<ConfigValidationIssue> {
         let mut issues = Vec::new();
@@ -6452,18 +6171,6 @@ fn default_imessage_bridge_url_env() -> Option<String> {
 
 fn default_imessage_bridge_token_env() -> Option<String> {
     Some(IMESSAGE_BRIDGE_TOKEN_ENV.to_owned())
-}
-
-fn default_signal_service_url() -> String {
-    "http://127.0.0.1:8080".to_owned()
-}
-
-fn default_signal_account_env() -> Option<String> {
-    Some(SIGNAL_ACCOUNT_ENV.to_owned())
-}
-
-fn default_signal_service_url_env() -> Option<String> {
-    Some(SIGNAL_SERVICE_URL_ENV.to_owned())
 }
 
 fn default_slack_api_base_url() -> String {
