+++ crates/app/src/config/mod.rs
@@ -22,21 +22,20 @@ pub use channels::{
     ImessageChannelConfig, IrcAccountConfig, IrcChannelConfig, LineAccountConfig,
     LineChannelConfig, MatrixAccountConfig, MatrixChannelConfig, MattermostAccountConfig,
     MattermostChannelConfig, NextcloudTalkAccountConfig, NextcloudTalkChannelConfig,
-    NostrAccountConfig, NostrChannelConfig,
-    ResolvedDingtalkChannelConfig, ResolvedDiscordChannelConfig, ResolvedEmailChannelConfig,
-    ResolvedFeishuChannelConfig, ResolvedGoogleChatChannelConfig, ResolvedImessageChannelConfig,
-    ResolvedIrcChannelConfig, ResolvedLineChannelConfig, ResolvedMatrixChannelConfig,
-    ResolvedMattermostChannelConfig, ResolvedNextcloudTalkChannelConfig,
-    ResolvedNostrChannelConfig, ResolvedSignalChannelConfig, ResolvedSlackChannelConfig,
-    ResolvedSynologyChatChannelConfig, ResolvedTeamsChannelConfig, ResolvedTelegramChannelConfig,
-    ResolvedTwitchChannelConfig, ResolvedWebhookChannelConfig, ResolvedWecomChannelConfig,
-    ResolvedWhatsappChannelConfig,
-    SignalAccountConfig, SignalChannelConfig, SlackAccountConfig, SlackChannelConfig,
-    SynologyChatAccountConfig, SynologyChatChannelConfig, TeamsAccountConfig, TeamsChannelConfig,
-    TelegramAccountConfig, TelegramChannelConfig, TelegramStreamingMode, TwitchAccountConfig,
-    TwitchChannelConfig, WebhookAccountConfig, WebhookChannelConfig, WebhookPayloadFormat,
-    WecomAccountConfig, WecomChannelConfig, WhatsappAccountConfig, WhatsappChannelConfig,
-    channel_descriptor, service_channel_descriptors,
+    NostrAccountConfig, NostrChannelConfig, ResolvedDingtalkChannelConfig,
+    ResolvedDiscordChannelConfig, ResolvedEmailChannelConfig, ResolvedFeishuChannelConfig,
+    ResolvedGoogleChatChannelConfig, ResolvedImessageChannelConfig, ResolvedIrcChannelConfig,
+    ResolvedLineChannelConfig, ResolvedMatrixChannelConfig, ResolvedMattermostChannelConfig,
+    ResolvedNextcloudTalkChannelConfig, ResolvedNostrChannelConfig, ResolvedSignalChannelConfig,
+    ResolvedSlackChannelConfig, ResolvedSynologyChatChannelConfig, ResolvedTeamsChannelConfig,
+    ResolvedTelegramChannelConfig, ResolvedTwitchChannelConfig, ResolvedWebhookChannelConfig,
+    ResolvedWecomChannelConfig, ResolvedWhatsappChannelConfig, SignalAccountConfig,
+    SignalChannelConfig, SlackAccountConfig, SlackChannelConfig, SynologyChatAccountConfig,
+    SynologyChatChannelConfig, TeamsAccountConfig, TeamsChannelConfig, TelegramAccountConfig,
+    TelegramChannelConfig, TelegramStreamingMode, TwitchAccountConfig, TwitchChannelConfig,
+    WebhookAccountConfig, WebhookChannelConfig, WebhookPayloadFormat, WecomAccountConfig,
+    WecomChannelConfig, WhatsappAccountConfig, WhatsappChannelConfig, channel_descriptor,
+    service_channel_descriptors,
 };
 #[allow(unused_imports)]
 pub(crate) use channels::{
@@ -50,15 +49,11 @@ pub(crate) use channels::{
     NOSTR_RELAY_URLS_ENV, SIGNAL_ACCOUNT_ENV, SIGNAL_SERVICE_URL_ENV, SLACK_BOT_TOKEN_ENV,
     SYNOLOGY_CHAT_INCOMING_URL_ENV, SYNOLOGY_CHAT_TOKEN_ENV, TEAMS_APP_ID_ENV,
     TEAMS_APP_PASSWORD_ENV, TEAMS_TENANT_ID_ENV, TEAMS_WEBHOOK_URL_ENV, TELEGRAM_BOT_TOKEN_ENV,
-    WEBHOOK_AUTH_TOKEN_ENV, WEBHOOK_ENDPOINT_URL_ENV, WEBHOOK_SIGNING_SECRET_ENV, WECOM_BOT_ID_ENV,
-    WECOM_SECRET_ENV, WHATSAPP_ACCESS_TOKEN_ENV, WHATSAPP_APP_SECRET_ENV,
-    WHATSAPP_PHONE_NUMBER_ID_ENV, WHATSAPP_VERIFY_TOKEN_ENV, normalize_channel_account_id,
-    parse_email_smtp_endpoint, parse_nostr_private_key_hex, parse_nostr_public_key_hex,
     TWITCH_ACCESS_TOKEN_ENV, WEBHOOK_AUTH_TOKEN_ENV, WEBHOOK_ENDPOINT_URL_ENV,
-    WEBHOOK_SIGNING_SECRET_ENV, WECOM_BOT_ID_ENV, WECOM_SECRET_ENV,
-    WHATSAPP_ACCESS_TOKEN_ENV, WHATSAPP_APP_SECRET_ENV, WHATSAPP_PHONE_NUMBER_ID_ENV,
-    WHATSAPP_VERIFY_TOKEN_ENV, normalize_channel_account_id, parse_email_smtp_endpoint,
-    parse_nostr_private_key_hex, parse_nostr_public_key_hex,
+    WEBHOOK_SIGNING_SECRET_ENV, WECOM_BOT_ID_ENV, WECOM_SECRET_ENV, WHATSAPP_ACCESS_TOKEN_ENV,
+    WHATSAPP_APP_SECRET_ENV, WHATSAPP_PHONE_NUMBER_ID_ENV, WHATSAPP_VERIFY_TOKEN_ENV,
+    normalize_channel_account_id, parse_email_smtp_endpoint, parse_nostr_private_key_hex,
+    parse_nostr_public_key_hex,
 };
 #[allow(unused_imports)]
 pub use conversation::{ConversationConfig, ConversationTurnLoopConfig};
