+++ docs/releases/architecture-drift-2026-03.md
@@ -1,7 +1,7 @@
 # Architecture Drift Report 2026-03
 
 ## Summary
-- Generated at: 2026-03-27T15:58:50Z
+- Generated at: 2026-03-27T17:01:08Z
 - Report month: `2026-03`
 - Baseline report: none
 - Hotspots tracked: 14
@@ -18,18 +18,18 @@
 | memory_mod | `foundation` | `crates/app/src/memory/mod.rs` | 356 | 650 | 294 | 14 | 16 | 2 | 87.5% | WATCH |
 | acp_manager | `operational_density` | `crates/app/src/acp/manager.rs` | 3327 | 3600 | 273 | 8 | 12 | 4 | 92.4% | WATCH |
 | acpx_runtime | `operational_density` | `crates/app/src/acp/acpx.rs` | 2696 | 2800 | 104 | 56 | 65 | 9 | 96.3% | TIGHT |
-| channel_registry | `structural_size` | `crates/app/src/channel/registry.rs` | 10432 | 10500 | 68 | 90 | 90 | 0 | 100.0% | TIGHT |
-| channel_config | `structural_size` | `crates/app/src/config/channels.rs` | 9490 | 9800 | 310 | 90 | 90 | 0 | 100.0% | TIGHT |
+| channel_registry | `structural_size` | `crates/app/src/channel/registry.rs` | 10439 | 10500 | 61 | 90 | 90 | 0 | 100.0% | TIGHT |
+| channel_config | `structural_size` | `crates/app/src/config/channels.rs` | 9493 | 9800 | 307 | 90 | 90 | 0 | 100.0% | TIGHT |
 | chat_runtime | `structural_size,operational_density` | `crates/app/src/chat.rs` | 6821 | 7300 | 479 | 145 | 160 | 15 | 93.4% | WATCH |
-| channel_mod | `structural_size,operational_density` | `crates/app/src/channel/mod.rs` | 6395 | 6400 | 5 | 106 | 110 | 4 | 99.9% | TIGHT |
+| channel_mod | `structural_size,operational_density` | `crates/app/src/channel/mod.rs` | 6379 | 6400 | 21 | 104 | 110 | 6 | 99.7% | TIGHT |
 | turn_coordinator | `structural_size,operational_density` | `crates/app/src/conversation/turn_coordinator.rs` | 9964 | 11200 | 1236 | 92 | 120 | 28 | 89.0% | WATCH |
 | tools_mod | `structural_size` | `crates/app/src/tools/mod.rs` | 14199 | 15000 | 801 | 54 | 70 | 16 | 94.7% | WATCH |
-| daemon_lib | `structural_size` | `crates/daemon/src/lib.rs` | 5971 | 6000 | 29 | 189 | 190 | 1 | 99.5% | TIGHT |
+| daemon_lib | `structural_size` | `crates/daemon/src/lib.rs` | 5897 | 6000 | 103 | 190 | 190 | 0 | 100.0% | TIGHT |
 | onboard_cli | `structural_size` | `crates/daemon/src/onboard_cli.rs` | 9256 | 9800 | 544 | 227 | 250 | 23 | 94.4% | WATCH |
 
 ## Prioritization Signals
 - BREACH hotspots (>100% of any tracked budget): none
-- TIGHT hotspots (>=95% of any tracked budget): spec_runtime (95.4%), acpx_runtime (96.3%), channel_registry (100.0%), channel_config (100.0%), channel_mod (99.9%), daemon_lib (99.5%)
+- TIGHT hotspots (>=95% of any tracked budget): spec_runtime (95.4%), acpx_runtime (96.3%), channel_registry (100.0%), channel_config (100.0%), channel_mod (99.7%), daemon_lib (100.0%)
 - WATCH hotspots (>=85% and <95% of any tracked budget): memory_mod (87.5%), acp_manager (92.4%), chat_runtime (93.4%), turn_coordinator (89.0%), tools_mod (94.7%), onboard_cli (94.4%)
 - Mixed-class hotspots (size plus operational density): chat_runtime, channel_mod, turn_coordinator
 
@@ -63,13 +63,13 @@
 <!-- arch-hotspot key=memory_mod lines=356 functions=14 -->
 <!-- arch-hotspot key=acp_manager lines=3327 functions=8 -->
 <!-- arch-hotspot key=acpx_runtime lines=2696 functions=56 -->
-<!-- arch-hotspot key=channel_registry lines=10432 functions=90 -->
-<!-- arch-hotspot key=channel_config lines=9490 functions=90 -->
+<!-- arch-hotspot key=channel_registry lines=10439 functions=90 -->
+<!-- arch-hotspot key=channel_config lines=9493 functions=90 -->
 <!-- arch-hotspot key=chat_runtime lines=6821 functions=145 -->
-<!-- arch-hotspot key=channel_mod lines=6395 functions=106 -->
+<!-- arch-hotspot key=channel_mod lines=6379 functions=104 -->
 <!-- arch-hotspot key=turn_coordinator lines=9964 functions=92 -->
 <!-- arch-hotspot key=tools_mod lines=14199 functions=54 -->
-<!-- arch-hotspot key=daemon_lib lines=5971 functions=189 -->
+<!-- arch-hotspot key=daemon_lib lines=5897 functions=190 -->
 <!-- arch-hotspot key=onboard_cli lines=9256 functions=227 -->
 <!-- arch-boundary key=memory_literals status=PASS -->
 <!-- arch-boundary key=provider_mod_helper_definitions status=PASS -->
